Update config.guess to recognize aarch64
Tizen toolchain team in Samsung company is planning to build Tizen
for new arm 64 bit architecture - Aarch64. But package
"libmm-wfd" cannot be built for this architecture because
of the following build errors:

""
checking build system type... ./config.guess: unable to guess system
type

This script, last modified 2011-05-11, has failed to recognize
the operating system you are using. It is advised that you
download the most up to date version of the config scripts from

   http://git.savannah.gnu.org/gitweb/?p=config.git;a=blob_plain;f=config.guess;hb=HEAD
and
   http://git.savannah.gnu.org/gitweb/?p=config.git;a=blob_plain;f=config.sub;hb=HEAD

If the version you run (./config.guess) is already up to date, please
send the following data and any information you think might be
pertinent to <config-patches@gnu.org> in order to provide the needed
information to handle your system.
""

Thus we have downloaded the latest version of file "config.guess"
from the suggested place and applied it to the git repository.

Change-Id: If2630cc1a44eee09e887f6e22eabec3191f91a38
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>